kivy/requirements.txt
kivymd==1.2.0
kivy[base]
plyer
piper-tts
pyglet